fix: handle suggested connectors
use correct component
remove redundant hook
use maintenance status
